clean(MicrosoftSentinelIntel): Cleanup, rename and move
