+++ stream/microsoft-sentinel-intel/src/microsoft_sentinel_intel/__init__.py
@@ -1,0 +1,9 @@
+from microsoft_sentinel_intel.client import ConnectorClient
+from microsoft_sentinel_intel.config import ConnectorSettings
+from microsoft_sentinel_intel.connector import Connector
+
+__all__ = [
+    "Connector",
+    "ConnectorClient",
+    "ConnectorSettings",
+]
