+++ stream/microsoft-sentinel-intel/src/microsoft_sentinel_intel/errors.py
@@ -1,0 +1,4 @@
+class ConnectorClientError(Exception):
+    def __init__(self, message, metadata):
+        self.message = message
+        self.metadata = metadata
