+++ stream/microsoft-sentinel-intel/src/microsoft_sentinel_intel/utils.py
