--- stream/microsoft-sentinel-intel/src/microsoft_sentinel_intel_connector/__init__.py
@@ -1,3 +1,0 @@
-from .connector import MicrosoftSentinelIntelConnector
-
-__all__ = ["MicrosoftSentinelIntelConnector"]
